Revert "Remove code duplication by moving addtopicdraft method to add-child-unit-button component"
This reverts commit 980fafdf3859f8478ee33972a74f732da6016c1b.
