+++ frontend/src/app/okrview/company/company.component.ts
@@ -144,6 +144,29 @@ export class CompanyComponent implements OnInit {
     this.currentOkrViewService.refreshCurrentCompanyView(company.id);
   }
 
+  // TODO: (R.J. 17.02.21) Refactor this and other occurrence in OkrChildUnitTabComponent, to remove duplicate code.
+  clickedAddTopicDraft(company: CompanyUnit): void {
+    const dialogReference: MatDialogRef<TopicDraftCreationFormComponent> = this.matDialog.open(TopicDraftCreationFormComponent, {
+      width: '600px', data: { companyId: company.id }
+    });
+
+    dialogReference
+      .afterClosed()
+      .pipe(
+        take(1),
+        filter(v => v),
+        switchMap(n => n)
+      )
+      .subscribe(addedTopicDraft => {
+        const snackBarText: string = this.i18n({
+          id: 'snackbar_addTopicDraft',
+          value: 'Ihr Themenentwurf wurde zur Prüfung abgeschickt.'
+        });
+        const snackBarOk: string = this.i18n({ id: 'snackbar_ok', value: 'Ok' });
+        this.snackBar.open(snackBarText, snackBarOk, { verticalPosition: 'top' });
+      });
+  }
+
   clickedDownloadExcelFileForCompany(company: CompanyUnit): void {
     this.excelFileService.downloadExcelFileForCompany(company.id);
   }
