+++ frontend/src/app/okrview/okr-child-unit/okr-child-unit-tab/okr-child-unit-tab.component.html
@@ -8,9 +8,9 @@
       <div class="icon-spacer f-1">
         <app-add-child-unit-button [userRole]="currentUserRole"
                                    [cycleClosed]="cycle.isCycleClosed()"
-                                   [addTopicDraftID]="short_okay"
                                    (clickedAddSubBranch)="clickedAddChildOkrBranch()"
-                                   (clickedAddSubDepartment)="clickedAddChildDepartment()">
+                                   (clickedAddSubDepartment)="clickedAddChildDepartment()"
+                                   (clickedAddTopicDraft)="clickedAddTopicDraft()">
         </app-add-child-unit-button>
       </div>
     </div>
